requirements.txt
# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart
python-dotenv

# Machine Learning Core
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0


tqdm
holidays
xarray
netcdf4
cftime
# Scheduling
schedule

# AI Integration
openai>=1.0.0
google-generativeai>=0.3.0

# Async Operations
aiohttp
aiofiles

# Google Cloud Platform
google-cloud-storage>=2.10.0


# Validation
pydantic>=2.0.0
pydantic-settings

# HTTP Client
httpx

# Development Dependencies
pytest
pytest-asyncio
mypy
black
flake8